Extendimg BoardState, starting to determine valid moves

+++ src/ChessPlatform/ChessConstants.cs
@@ -8,13 +8,13 @@ namespace ChessPlatform
     {
         #region Constants and Fields
 
-        public const int FileCount = 8;
-        public const int RankCount = 8;
+        public const byte FileCount = 8;
+        public const byte RankCount = 8;
 
-        public const int WhitePawnPromotionRank = RankCount - 1;
-        public const int BlackPawnPromotionRank = 0;
+        public const byte WhitePawnPromotionRank = RankCount - 1;
+        public const byte BlackPawnPromotionRank = 0;
 
-        internal const int X88Length = FileCount * RankCount * 2;
+        internal const byte X88Length = FileCount * RankCount * 2;
 
         public static readonly ValueRange<int> FileRange = ValueRange.Create(0, FileCount - 1);
         public static readonly ValueRange<int> RankRange = ValueRange.Create(0, RankCount - 1);
